+++ src/wx/image_sequence_dialog.h
@@ -25,7 +25,7 @@ class ImageSequenceDialog : public TableDialog
 public:
 	ImageSequenceDialog (wxWindow* parent);
 
-	float frame_rate () const;
+	double frame_rate () const;
 
 private:
 	wxTextCtrl* _frame_rate;
